added next(err) code for req.session

+++ src/server/routes/excelTemplates/deleteTemplate.ts
@@ -1,48 +1,43 @@
-import express, { Request } from 'express';
-import appRootPath from 'app-root-path'
-import fs from 'fs';
-import path from 'path';
-
+import express, { Request } from "express";
+import appRootPath from "app-root-path";
+import fs from "fs";
+import path from "path";
 
 const router = express.Router();
 
 interface session {
-    login: boolean,
-    uID: number,
+    login: boolean;
+    uID: number;
 }
 
 interface uploadTemplatePost extends Request {
-    session: session,
-    files: any
+    session: session;
+    files: any;
 }
 
 interface resObj {
-    message: Text,
+    message: Text;
 }
 
-//fs.unlink
-//fs.existsync
-
-router.get('/deleteTemplate', (req: uploadTemplatePost, res: any) => {
+//@ts-ignore
+router.get("/deleteTemplate", (req: uploadTemplatePost, res: any, next) => {
     try {
+        if (req.session === undefined) throw new Error("Login request not associated with session.");
         if (req.session.login === true) {
-            const template = req.query['template'];
-            const userDirecotry: string = `${appRootPath}/uploads/${req.session.uID}/`
-            const path: string = `${userDirecotry}${template}`
-            console.log('------------------------', template, userDirecotry, path)
-
+            const template = req.query["template"];
+            const userDirecotry: string = `${appRootPath}/uploads/${req.session.uID}/`;
+            const path: string = `${userDirecotry}${template}`;
             if (fs.existsSync(path)) {
-                console.log('path exists:', path)
-                fs.unlinkSync(path)
-                res.status(200).send({ message: 'File deleted' })
+                console.log("path exists:", path);
+                fs.unlinkSync(path);
+                res.status(200).send({ message: "File deleted" });
             } else {
-                res.status(500).send({ message: 'could not find file.' });
+                res.status(500).send({ message: "could not find file." });
             }
         }
-    } catch (err) {
-        console.log(err)
-        res.status(500).send({ message: err });
+    } catch (error) {
+        next(error);
     }
-})
+});
 
-export default router
+export default router;
